NextRole/requirements.txt
streamlit==1.43.2
pdfplumber==0.11.5
python-docx==1.1.2
openai==1.75.0
python-dotenv==1.0.1
tiktoken>=0.5.0
